implement catalog scan segment progress
